move `load_one_graph` and `get`  to Trainer
